[runc-shim] Move async code to package root
